comparision_service/requirements.txt
connexion>=2, <3
connexion[swagger-ui]>=2, <3
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
beautifulsoup4==4.12.3
bs4==0.0.2 
soupsieve==2.5
python-dotenv==1.0.1
async-timeout==4.0.3 
redis==5.0.4
Flask-SocketIO==5.3.6 
bidict==0.23.1 
h11==0.14.0 
python-engineio==4.9.0 
python-socketio==5.11.2 
simple-websocket==1.0.0 
wsproto==1.2.0
kafka-python==2.0.2
requests==2.31.0
flask_cors==4.0.1